#1351 created abstract class and cleaned code
